test: Generalize benchmark variants resolution
